Refactor note handling and add Postman access for registration

+++ controllers/noteControllers.js
@@ -1,6 +1,5 @@
 const { responseReturn } = require("../utils/response");
 const noteModel = require("../models/noteModel");
-const { Types } = require('mongoose');
 const userModel = require("../models/userModel");
 const delay = require("../utils/delay");
 
@@ -8,48 +7,46 @@ class noteControllers {
     add_note = async (req, res) => {
         await delay(200);
         const {id} = req;
-        const {title, content} = req.body;
-        if (!title || !content) {
-            return responseReturn(res, 400, {error: 'Please fill all fields!'})
-        } else {
-            try {
-                const newNote = await noteModel.create({
-                    id: new Types.ObjectId(),
-                    userId: id,
-                    title,
-                    content
-                })
-                await newNote.save();
-                await userModel.findByIdAndUpdate(id, { $inc: {noteTotal: 1} });
-                responseReturn(res, 200, {message: 'Note added!', note: newNote})
-            } catch (error) {
-                responseReturn(res, 500, {error: 'Internal Server Error!'})
-            }
-        }        
+        const {title, content} = req.body;        
+        try {
+            const newNote = await noteModel.create({
+                userId: id,
+                title,
+                content
+            })
+            await userModel.findByIdAndUpdate(id, { $inc: {noteTotal: 1} });
+            responseReturn(res, 201, {message: 'Note added!', note: newNote})
+        } catch (error) {
+            responseReturn(res, 500, {error: 'Internal Server Error!'})
+        }
     }
     // End method
     get_notes = async (req, res) => {
         const {id} = req;
         try {
             const notes = await noteModel.find({userId: id});
-            const totalNotes = await noteModel.find({userId: id}).countDocuments();
+            const totalNotes = notes.length;
             responseReturn(res, 200, {notes, totalNotes})
         } catch (error) {
-            console.log(error.message);
+            responseReturn(res, 500, {error: 'Internal Server Error!'})
         }
     }
     // End method
     delete_note = async (req, res) => {
         try {
             const {id} = req.params;
-            await noteModel.findByIdAndDelete(id);
-            if (!id) {
-                return res.status(400).json({error: 'Note not found!'})
+            const note = await noteModel.findById(id);
+            if (!note) {
+                return responseReturn(res, 404, {error: 'Note not found!'})
             }
+            if (note.userId.toString() !== req.id) {
+                return responseReturn(res, 403, {error: 'Forbidden!'})
+            }
+            await noteModel.findByIdAndDelete(id);
             await userModel.findByIdAndUpdate(req.id, { $inc: {noteTotal: -1} });
-            res.status(200).json({message: 'Note deleted!'})
+            responseReturn(res, 200, {message: 'Note deleted!'})
         } catch (error) {
-            res.status(500).json({error: 'Internal Server Error!'})
+            responseReturn(res, 500, {error: 'Internal Server Error!'})
         }
 
     }
